requirements-avatar.txt
torch==1.13.1
torchvision==0.14.1
torchaudio==0.13.1
transformers==4.21.3
pillow==9.5.0
numpy==1.22.0
opencv-python==4.8.0.76
scipy==1.10.1
TTS==0.15.6
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
# SadTalker dependencies
facelib==0.3.0
basicsr==1.4.2
realesrgan==0.3.0
gfpgan==1.3.8
yacs==0.1.8
dlib==19.24.2